refactor: Error 메시지 InputView, Lotto 클래스별로 분리 작성
