+++ src/main/java/lotto/controller/LottoController.java
@@ -2,6 +2,7 @@ package lotto.controller;
 
 import java.util.List;
 import lotto.domain.BonusNumber;
+import lotto.domain.Lotto;
 import lotto.utils.LottoExceptionUtils;
 import lotto.view.InputView;
 
@@ -10,9 +11,8 @@ public class LottoController {
     public void runLotto() {
         int lottoPurchaseAmount = LottoExceptionUtils.runUntilNoneLottoException(
                 InputView::inputLottoPurchaseAmount);
-        String wonLottoNumbersString = LottoExceptionUtils.runUntilNoneLottoException(
-                InputView::inputWonLottoNumbers);
-        List<Integer> parsedWonLottoNumbers = InputView.parseWonLottoNumbers(wonLottoNumbersString);
+        Lotto wonLotto = LottoExceptionUtils.runUntilNoneLottoException(
+                InputView::wonLottoNumbers);
         BonusNumber bonusNumber = LottoExceptionUtils.runUntilNoneLottoException(InputView::inputBonusNumber);
     }
 }
